changues: resolve bug

+++ src/db/migrations/20231019175134-migrating-permissions-customer.ts
@@ -6,32 +6,38 @@ const { PERMISSION_TABLE } = permissionModel;
 export async function up(queryInterface: QueryInterface) {
   const deleteCriteria = {
     code: {
-      [Op.and]: [{ [Op.startsWith]: "P03" }, { [Op.notLike]: "P03-%" }],
+      [Op.startsWith]: "P03-",
     },
   };
 
   const newPermissions = [
     {
-      name: "COMENTARIOS",
+      name: "DETALLES - CLIENTE",
       code: "P02-02",
       icon: "-",
       link: "/cobranza/:urlIdentifier/clientes/:code",
     },
     {
-      name: "AGREGAR COMENTARIO",
+      name: "COMENTARIOS",
       code: "P02-02-01",
+      icon: "-",
+      link: "/cobranza/:urlIdentifier/clientes/:code/gestion",
+    },
+    {
+      name: "AGREGAR COMENTARIO",
+      code: "P02-02-01-01",
       icon: "-",
       link: "#",
     },
     {
       name: "ACTUALIZAR COMENTARIO",
-      code: "P02-02-02",
+      code: "P02-02-01-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR COMENTARIO",
-      code: "P02-02-03",
+      code: "P02-02-01-03",
       icon: "-",
       link: "#",
     },
@@ -55,114 +61,112 @@ export async function up(queryInterface: QueryInterface) {
     },
     {
       name: "WORD - DESCARGAR INFORMACIÓN DEL CLIENTE",
-      code: "P02-06",
+      code: "P02-02-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ARCHIVOS",
-      code: "P02-07",
+      code: "P02-02-03",
       icon: "-",
       link: "#",
     },
     {
       name: "VER ARCHIVO",
-      code: "P02-07-01",
+      code: "P02-02-03-01",
       icon: "-",
       link: "#",
     },
     {
       name: "AGREGAR ARCHIVO",
-      code: "P02-07-02",
+      code: "P02-02-03-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR ARCHIVO",
-      code: "P02-07-03",
+      code: "P02-02-03-03",
       icon: "-",
       link: "#",
     },
     {
       name: "FIADORES",
-      code: "P02-08",
+      code: "P02-02-04",
       icon: "-",
       link: "#",
     },
     {
       name: "AGREGAR FIADOR",
-      code: "P02-08-01",
+      code: "P02-02-04-01",
       icon: "-",
       link: "#",
     },
     {
       name: "ACTUALIZAR FIADOR",
-      code: "P02-08-02",
+      code: "P02-02-04-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR FIADOR",
-      code: "P02-08-03",
+      code: "P02-02-04-03",
       icon: "-",
       link: "#",
     },
     {
       name: "DIRECCIONES",
-      code: "P02-09",
+      code: "P02-02-05",
       icon: "-",
       link: "#",
     },
     {
       name: "AGREGAR DIRECCIÓN",
-      code: "P02-09-01",
+      code: "P02-02-05-01",
       icon: "-",
       link: "#",
     },
     {
       name: "ACTUALIZAR DIRECCIÓN",
-      code: "P02-09-02",
+      code: "P02-02-05-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR DIRECCIÓN",
-      code: "P02-09-03",
+      code: "P02-02-05-03",
       icon: "-",
       link: "#",
     },
     {
       name: "PRODUCTOS",
-      code: "P02-10",
+      code: "P02-02-06",
       icon: "-",
       link: "#",
     },
     {
       name: "AGREGAR PRODUCTO",
-      code: "P02-10-01",
+      code: "P02-02-06-01",
       icon: "-",
       link: "#",
     },
     {
       name: "ACTUALIZAR PRODUCTO",
-      code: "P02-10-02",
+      code: "P02-02-06-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR PRODUCTO",
-      code: "P02-10-03",
+      code: "P02-02-06-03",
       icon: "-",
       link: "#",
     },
   ];
 
   return queryInterface.sequelize.transaction(async (transaction) => {
-    await queryInterface.bulkDelete(
-      PERMISSION_TABLE,
-      { deleteCriteria },
-      { transaction }
-    );
+    await queryInterface.bulkDelete(PERMISSION_TABLE, deleteCriteria, {
+      transaction,
+    });
     await queryInterface.bulkInsert(PERMISSION_TABLE, newPermissions, {
       transaction,
     });
@@ -172,9 +176,11 @@ export async function up(queryInterface: QueryInterface) {
 export async function down(queryInterface: QueryInterface) {
   const deleteCriteria = {
     code: {
-      [Op.and]: [{ [Op.startsWith]: "P02" }, { [Op.notLike]: "P02-%" }],
+      [Op.startsWith]: "P02-",
+      [Op.not]: "P02-01",
     },
   };
+
   const addPermissions = [
     {
       name: "COMENTARIOS",
@@ -183,26 +189,26 @@ export async function down(queryInterface: QueryInterface) {
       link: "/cobranza/:urlIdentifier/cobranza/:code",
     },
     {
-      name: "AGREGAR CLIENTE",
-      code: "P03-02",
-      icon: "-",
-      link: "#",
-    },
-    {
       name: "AGREGAR COMENTARIO",
-      code: "P03-03-01",
+      code: "P03-01-01",
       icon: "-",
       link: "#",
     },
     {
       name: "ACTUALIZAR COMENTARIO",
-      code: "P03-03-02",
+      code: "P03-01-02",
       icon: "-",
       link: "#",
     },
     {
       name: "ELIMINAR COMENTARIO",
-      code: "P03-03-03",
+      code: "P03-01-03",
+      icon: "-",
+      link: "#",
+    },
+    {
+      name: "AGREGAR CLIENTE",
+      code: "P03-02",
       icon: "-",
       link: "#",
     },
@@ -323,11 +329,9 @@ export async function down(queryInterface: QueryInterface) {
   ];
 
   return queryInterface.sequelize.transaction(async (transaction) => {
-    await queryInterface.bulkDelete(
-      PERMISSION_TABLE,
-      { deleteCriteria },
-      { transaction }
-    );
+    await queryInterface.bulkDelete(PERMISSION_TABLE, deleteCriteria, {
+      transaction,
+    });
     await queryInterface.bulkInsert(PERMISSION_TABLE, addPermissions, {
       transaction,
     });
